Fixed build & Added some more tests.
